faster-rcnn/requirements.txt
# PyTorch and CUDA
# Install PyTorch from: https://pytorch.org/get-started/locally/
# Example: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# Detectron2
# Install from: https://detectron2.readthedocs.io/en/latest/tutorials/install.html
# git clone https://github.com/facebookresearch/detectron2.git
# cd detectron2 && pip install -e .

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Training Tools
tensorboard>=2.13.0
tqdm>=4.65.0

# Dataset and Annotations
pycocotools>=2.0.6

# Data Augmentation
albumentations>=1.3.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities
pyyaml>=6.0
